+++ release-notes.md
@@ -6,7 +6,7 @@ for [Chrome extension to for Bible.com](README.md).
 
 ## 2.2.0 (🔜 coming soon)
 
-- [ ] ...
+- [x] Don't allow to remove images from settings if they are used in slides
 
 ## 2.1.0